move Settings modals into own directory
