+++ src/components/displays/OscilloscopeChart.tsx
@@ -5,7 +5,7 @@ import 'chartjs-adapter-luxon';
 import { useBackend } from "@/backend_interop/BackendProvider";
 import { unDecimatedPackets } from "@/backend_interop/buffers";
 import { PacketComponentType, PacketField } from "@/backend_interop/types";
-import { OscilloscopeGraphStruct } from "@/modals/OscilloscopeGraphSettingsModal";
+import { OscilloscopeGraphStruct } from "@/modals/settings/OscilloscopeGraphSettingsModal";
 
 // Register the necessary components with ChartJS so that they can be used later
 // Note: any components that are not registered here will act like no-ops if they are attempted to be used later!
