+++ src/components/displays/SolidChart.tsx
@@ -5,7 +5,7 @@ import 'chartjs-adapter-luxon';
 import { useBackend } from "@/backend_interop/BackendProvider";
 import { unDecimatedPackets } from "@/backend_interop/buffers";
 import { PacketComponentType, PacketField } from "@/backend_interop/types";
-import { GraphStruct } from "@/modals/GraphSettingsModal";
+import { GraphStruct } from "@/modals/settings/GraphSettingsModal";
 
 // Register the necessary components with ChartJS so that they can be used later
 // Note: any components that are not registered here will act like no-ops if they are attempted to be used later!
