+++ src/modals/settings/OscilloscopeGraphSettingsModal.tsx
